Avoid compiler warnings about possible loss of data
This quenches the compiler warning of MSVC á la:

	conversion from 'int64_t' to 'int', possible loss of data

+++ src/win32/mouse.c
@@ -26,7 +26,7 @@
  */
 void moveMouse(MMPoint point)
 {
-	SetCursorPos (point.x, point.y);
+	SetCursorPos ((int)point.x, (int)point.y);
 }
 
 void dragMouse(MMPoint point, const MMMouseButton button)
